gr-uhd: evaluate stream_chans to generate channels
In the yml blocks for uhd src and sink, the channels parameter was not
being populated correctly, as the %if to tell the template how to render
the make function evaluate the string representation of stream_chans,
looking for whether it was None.

Perhaps there is a cleaner way to do this, but this allows stream_chans
to be set to [] in the grc parameters and still evaluate as None so that
range(N) is rendered for a multichannel usrp.